Add parallel registration and simultaneous fitting capabilities (incomplete)
Modified the ModelManagement script to allow parallel registration by distributing various redshift points to seperate processes.
Modified ModelManagement and Preperation scripts to allow simultaneous fitting using ```filter_scaling``` attribute. This is a quick and dirty implementation that is going to be improved in the next version.
This commit is likely incomplete, pending a final test using these scripts.